chore: renamed file to map_overlay_line_style.dart
